requirements.txt
fastapi==0.128.0
pandas==3.0.0
python-dotenv==1.2.1
Requests==2.32.5
starlette==0.50.0
uvicorn==0.40.0
openpyxl==3.1.5
Jinja2==3.1.6
python-multipart==0.0.22
itsdangerous==2.2.0
bcrypt==5.0.0
httpx==0.28.1
pytest==8.4.2
